@@ -1,4 +1,4 @@
-name: freeclimbers_employee
+name: climbers
 description: Climbers App
 
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
@@ -15,7 +15,7 @@ dependencies:
     sdk: flutter
 
   cupertino_icons: ^1.0.2
-  go_router: ^6.2.0
+  go_router: ^14.2.0
   uni_links: ^0.5.1
   oktoast: ^3.3.1
   dotted_border: ^2.0.0+3
@@ -23,23 +23,23 @@ dependencies:
     git:
       url: https://github.com/Majestr32/flutter_adaptive_action_sheet.git
   flutter_native_splash: ^2.2.14
-  flutter_launcher_icons: ^0.11.0
-  font_awesome_flutter: 9.2.0
+  flutter_launcher_icons: ^0.13.1
+  font_awesome_flutter: ^10.7.0
   loading_indicator: ^3.1.0
-  app_links: ^3.4.3
-  loader_overlay: ^2.1.0
-  webview_flutter: ^3.0.4
-  image_picker: ^0.8.6
-  google_fonts: ^4.0.3
+  app_links: ^6.1.4
+  loader_overlay: ^4.0.0
+  webview_flutter: ^4.8.0
+  image_picker: ^1.1.2
+  google_fonts: ^6.2.1
   flutter_slidable:
     git:
       url: https://github.com/Majestr32/flutter_slidable_with_slide_callback.git
-  onesignal_flutter: ^3.5.0
+  onesignal_flutter: ^5.2.2
   carousel_slider: ^4.2.1
   local_auth: ^2.1.6
   universal_io: ^2.2.2
   redux_persist_flutter: ^0.9.0
-  image_cropper: ^3.0.1
+  image_cropper: ^7.1.0
   flutter_wallet:
     git: https://github.com/enolcasielles/flutter_wallet.git
   freezed: ^2.2.1
@@ -48,26 +48,26 @@ dependencies:
   path_provider: ^2.0.11
   device_display_brightness:
     git: https://github.com/ViacheslavSomin/device_display_brightness.git
-  screenshot: ^2.1.0
-  permission_handler: ^10.2.0
+  screenshot: ^3.0.0
+  permission_handler: ^11.3.1
   json_annotation: ^4.7.0
   freezed_annotation: ^2.2.0
   json_serializable: ^6.5.4
-  dio: ^4.0.6
+  dio: ^5.5.0+1
   bloc: ^8.1.0
   flutter_bloc: ^8.1.1
   equatable: ^2.0.5
   shared_preferences: ^2.0.15
   dart_jsonwebtoken: ^2.4.2
-  cupertino_http: ^0.0.10
-  http: ^0.13.5
+  cupertino_http: ^1.5.0
+  http: ^1.2.1
   super_rich_text: ^1.0.0
   url_launcher: ^6.1.10
   crypto: ^3.0.0
-  intl: ^0.18.0
+  intl: ^0.19.0
   qr: ^3.0.1
-  connectivity_plus: ^3.0.3
-  flutter_oss_licenses: ^2.0.1
+  connectivity_plus: ^6.0.3
+  flutter_oss_licenses: ^3.0.2
   overlay_tutorial: ^3.0.0
 
 flutter_native_splash:
@@ -85,7 +85,7 @@ dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  flutter_lints: ^2.0.0
+  flutter_lints: ^4.0.0
   build_runner: ^2.3.2
 
 flutter:
