@@ -2,20 +2,24 @@ import 'dart:async';
 import 'dart:convert';
 import 'dart:developer';
 import 'dart:typed_data';
-import 'package:freeclimbers_employee/blocs/connectivity_bloc/connectivity_bloc.dart';
+import 'package:climbers/blocs/connectivity_bloc/connectivity_bloc.dart';
+import 'package:climbers/blocs/member_cubit/member_cubit.dart';
 
 import 'package:bloc/bloc.dart';
-import 'package:freeclimbers_employee/models/app_documents/app_documents.dart';
-import 'package:freeclimbers_employee/models/app_settings/app_settings.dart';
-import 'package:freeclimbers_employee/repositories/app/app_repository.dart';
+import 'package:climbers/models/app_documents/app_documents.dart';
+import 'package:climbers/models/app_settings/app_settings.dart';
+import 'package:climbers/repositories/app/app_repository.dart';
 import 'package:equatable/equatable.dart';
 
 import '../../custom_error.dart';
+import '../branches_cubit/branches_cubit.dart';
 
 part 'app_state.dart';
 
 class AppCubit extends Cubit<AppState> {
 
+  final BranchesCubit _branchesCubit;
+  final MemberCubit _memberCubit;
   final AppRepository _appRepository;
   final ConnectivityBloc _connectivityBloc;
   final int? defaultBranch;
@@ -49,10 +53,15 @@ class AppCubit extends Cubit<AppState> {
     final previousState = state;
     try{
       int? selectedBranchFromShared = _appRepository.getSelectedBranchId() == null ? null : int.parse(_appRepository.getSelectedBranchId()!);
+      bool branchExists = _branchesCubit.state.branches.any((element) => element.details.no == selectedBranchFromShared.toString());
+      if(!branchExists){
+        selectedBranchFromShared = null;
+      }
+      int? actualSelectedBranch = branchExists ? selectedBranchFromShared : null;
 
-      final settings = await _appRepository.getBranchSettings(branchId: selectedBranchFromShared);
-      final documents = await _appRepository.getTermsAndServices(branchId: selectedBranchFromShared);
-      final logo = await _appRepository.getBase64Logo(branchId: selectedBranchFromShared);
+      final settings = await _appRepository.getBranchSettings(branchId: actualSelectedBranch);
+      final documents = await _appRepository.getTermsAndServices(branchId: actualSelectedBranch);
+      final logo = await _appRepository.getBase64Logo(branchId: actualSelectedBranch);
       emit(AppState.initial().copyWith(settings: settings, selectedBranchId: selectedBranchFromShared, documents: documents, branchIcon: logo == null ? null : base64Decode(logo), status: AppStateStatus.loaded));
     }on CustomError catch(e){
       emit(state.copyWith(status: AppStateStatus.error, error: e));
@@ -75,10 +84,14 @@ class AppCubit extends Cubit<AppState> {
 
 
   AppCubit({
+    required BranchesCubit branchesCubit,
+    required MemberCubit memberCubit,
     required AppRepository appRepository,
     required ConnectivityBloc connectivityBloc,
     this.defaultBranch
-  })  : _connectivityBloc = connectivityBloc,
+  })  : _branchesCubit = branchesCubit,
+        _memberCubit = memberCubit,
+        _connectivityBloc = connectivityBloc,
         _appRepository = appRepository, super(AppState.initial()){
     _connectivityBlocStreamSubscription = _connectivityBloc.stream.listen((event) {
       event.maybeMap(
@@ -88,6 +101,18 @@ class AppCubit extends Cubit<AppState> {
             }
           },
           orElse: (){});
+    });
+    _branchesCubitStreamSubscription = _branchesCubit.stream.listen((event) async{
+      await loadAppSettings();
+    });
+    _memberCubitStreamSubscription = _memberCubit.stream.listen((state) async{
+      if(state.status == MemberStateStatus.unauthenticated){
+        if(this.state.selectedBranchId == 0){
+          return;
+        }
+        await _appRepository.setSelectedBranchId("0");
+        loadAppSettings();
+      }
     });
   }
 
