@@ -1,8 +1,8 @@
-import 'package:freeclimbers_employee/blocs/app_cubit/app_cubit.dart';
-import 'package:freeclimbers_employee/blocs/theme_bloc/theme_bloc.dart';
-import 'package:freeclimbers_employee/consts/no_internet_theme_colors.dart';
-import 'package:freeclimbers_employee/models/color_set/color_set.dart';
-import 'package:freeclimbers_employee/models/theme_colors/theme_colors.dart';
+import 'package:climbers/blocs/app_cubit/app_cubit.dart';
+import 'package:climbers/blocs/theme_bloc/theme_bloc.dart';
+import 'package:climbers/consts/no_internet_theme_colors.dart';
+import 'package:climbers/models/color_set/color_set.dart';
+import 'package:climbers/models/theme_colors/theme_colors.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/scheduler.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
