@@ -1,38 +1,44 @@
 import 'dart:async';
 import 'dart:developer';
-import 'package:freeclimbers_employee/repositories/app/app_repository.dart';
-import 'package:freeclimbers_employee/repositories/branches/branches_repository.dart';
-import 'package:freeclimbers_employee/repositories/language/language_repository.dart';
-import 'package:freeclimbers_employee/repositories/member/member_repository.dart';
-import 'package:freeclimbers_employee/repositories/membercard/membercard_repository.dart';
-import 'package:freeclimbers_employee/repositories/qr_code/qr_code_repository.dart';
-import 'package:freeclimbers_employee/repositories/security/security_repository.dart';
-import 'package:freeclimbers_employee/repositories/theme_repository/theme_repository.dart';
-import 'package:freeclimbers_employee/repositories/tutorial/tutorial_repository.dart';
-import 'package:freeclimbers_employee/services/app/app_service.dart';
-import 'package:freeclimbers_employee/services/branches/branches_service.dart';
-import 'package:freeclimbers_employee/services/member/member_service.dart';
-import 'package:freeclimbers_employee/services/membercard/membercard_service.dart';
-import 'package:freeclimbers_employee/services/qr_code/qr_code_service.dart';
-import 'package:freeclimbers_employee/simple_bloc_observer.dart';
+import 'package:climbers/blocs/servers_bloc/servers_bloc.dart';
+import 'package:climbers/repositories/app/app_repository.dart';
+import 'package:climbers/repositories/branches/branches_repository.dart';
+import 'package:climbers/repositories/language/language_repository.dart';
+import 'package:climbers/repositories/member/member_repository.dart';
+import 'package:climbers/repositories/membercard/membercard_repository.dart';
+import 'package:climbers/repositories/qr_code/qr_code_repository.dart';
+import 'package:climbers/repositories/security/security_repository.dart';
+import 'package:climbers/repositories/servers/servers_repository.dart';
+import 'package:climbers/repositories/theme_repository/theme_repository.dart';
+import 'package:climbers/repositories/tutorial/tutorial_repository.dart';
+import 'package:climbers/services/app/app_service.dart';
+import 'package:climbers/services/branches/branches_service.dart';
+import 'package:climbers/services/member/member_service.dart';
+import 'package:climbers/services/membercard/membercard_service.dart';
+import 'package:climbers/services/qr_code/qr_code_service.dart';
+import 'package:climbers/simple_bloc_observer.dart';
+import 'package:climbers/ui/screens/auth/create_server.dart';
 import 'package:connectivity_plus/connectivity_plus.dart';
 import 'package:flutter/services.dart';
 import 'package:app_links/app_links.dart';
 
-import 'package:freeclimbers_employee/blocs/app_cubit/app_cubit.dart';
-import 'package:freeclimbers_employee/blocs/connectivity_bloc/connectivity_bloc.dart';
-import 'package:freeclimbers_employee/blocs/locale_cubit/locale_cubit.dart';
-import 'package:freeclimbers_employee/blocs/security_bloc/security_bloc.dart';
-import 'package:freeclimbers_employee/blocs/theme_bloc/theme_bloc.dart';
-import 'package:freeclimbers_employee/consts/k_colors.dart';
-import 'package:freeclimbers_employee/consts/no_internet_theme_colors.dart';
-import 'package:freeclimbers_employee/extensions/brightness.dart';
-import 'package:freeclimbers_employee/router.dart';
-import 'package:freeclimbers_employee/ui/screens/auth/sign_in.dart';
-import 'package:freeclimbers_employee/ui/screens/splash/splash_screen.dart';
-import 'package:freeclimbers_employee/utils/dio.dart';
-import 'package:freeclimbers_employee/ui/widgets/loading/overlay_loading.dart';
-import 'package:freeclimbers_employee/utils/toasts.dart';
+import 'package:climbers/blocs/app_cubit/app_cubit.dart';
+import 'package:climbers/blocs/connectivity_bloc/connectivity_bloc.dart';
+import 'package:climbers/blocs/locale_cubit/locale_cubit.dart';
+import 'package:climbers/blocs/member_cubit/member_cubit.dart';
+import 'package:climbers/blocs/security_bloc/security_bloc.dart';
+import 'package:climbers/blocs/theme_bloc/theme_bloc.dart';
+import 'package:climbers/consts/k_colors.dart';
+import 'package:climbers/consts/no_internet_theme_colors.dart';
+import 'package:climbers/extensions/brightness.dart';
+import 'package:climbers/router.dart';
+import 'package:climbers/ui/screens/auth/sign_in_email.dart';
+import 'package:climbers/ui/screens/splash/splash_screen.dart';
+import 'package:climbers/ui/widgets/app_lifestyle_overlay/app_lifestyle_overlay.dart';
+import 'package:climbers/ui/widgets/lock_screen_overlay/lock_screen_overlay.dart';
+import 'package:climbers/utils/dio.dart';
+import 'package:climbers/ui/widgets/loading/overlay_loading.dart';
+import 'package:climbers/utils/toasts.dart';
 import 'package:dio/dio.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
@@ -46,6 +52,10 @@ import 'package:oktoast/oktoast.dart';
 import 'package:onesignal_flutter/onesignal_flutter.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:shared_preferences/shared_preferences.dart';
+import 'package:uni_links/uni_links.dart';
+
+import 'blocs/branches_cubit/branches_cubit.dart';
+import 'blocs/tutorial_bloc/tutorial_bloc.dart';
 import 'data/shared_prefs/app_shared_prefs.dart';
 import 'l10n/l10n.dart';
 
@@ -60,8 +70,8 @@ void main() async{
   await sharedPrefs.reload();
   final appSharedPrefs = AppSharedPrefs(sharedPreferences: sharedPrefs);
 
-  OneSignal.shared.setLogLevel(OSLogLevel.verbose, OSLogLevel.none);
-  OneSignal.shared.setAppId('77d0c650-b9e8-4563-acec-5488ac8f30cf');
+  OneSignal.Debug.setLogLevel(OSLogLevel.verbose);
+  OneSignal.initialize('77d0c650-b9e8-4563-acec-5488ac8f30cf');
   
   Bloc.observer = SimpleBlocObserver();
 
@@ -72,18 +82,24 @@ void main() async{
     ..interceptors.add(LanguageInterceptor(localeCubit: localeCubit));
 
 
-  final defaultBranch = (await AppLinks().getInitialAppLink())?.queryParameters["bid"];
+  final defaultBranch = (await AppLinks().getInitialLink())?.queryParameters["bid"];
 
   final memberRepository = MemberRepository(memberService: MemberService(dio: dio), appSharedPrefs: appSharedPrefs);
   final memberCardRepository = MemberCardRepository(memberCardService: MemberCardService());
+  final memberCubit = MemberCubit(memberRepository: memberRepository, memberCardRepository: memberCardRepository);
   final rootLocalization = await AppLocalizations.delegate.load(Locale(localeCubit.state));
   final branchesRepository = BranchRepository(appService: AppService(dio: dio), branchService: BranchesService(dio: dio));
   final appRepository = AppRepository(appService: AppService(dio: dio), branchService: BranchesService(dio: dio), sharedPrefs: appSharedPrefs);
+  final branchCubit = BranchesCubit(branchesRepository: branchesRepository, sharedPrefs: appSharedPrefs, appRepository: appRepository, memberCubit: memberCubit);
   final connectivityBloc = ConnectivityBloc(connectionChecker: Connectivity());
-  final appCubit = AppCubit(appRepository: appRepository, defaultBranch: defaultBranch == null ? null : int.tryParse(defaultBranch), connectivityBloc: connectivityBloc);
+  final appCubit = AppCubit(appRepository: appRepository, defaultBranch: defaultBranch == null ? null : int.tryParse(defaultBranch), branchesCubit: branchCubit, connectivityBloc: connectivityBloc, memberCubit: memberCubit);
   final themeRepository = ThemeRepository(appSharedPrefs: appSharedPrefs);
   final themeBloc = ThemeBloc(themeRepository: themeRepository);
   final securityRepository = SecurityRepository(auth: LocalAuthentication(), localization: rootLocalization, appSharedPrefs: appSharedPrefs);
+  final securityBloc = SecurityBloc(securityRepository: securityRepository, memberCubit: memberCubit);
+  final serversRepository = ServersRepository(sharedPrefs: appSharedPrefs);
+  final serversBloc = ServersBloc(serversRepository: serversRepository);
+  serversBloc.add(const ServersEvent.fetchServers());
   
   runApp(MultiRepositoryProvider(
       providers: [
@@ -96,16 +112,22 @@ void main() async{
         RepositoryProvider(create: (_) => securityRepository),
         RepositoryProvider(create: (_) => TutorialRepository(appSharedPrefs: appSharedPrefs)),
         RepositoryProvider(create: (_) => ThemeRepository(appSharedPrefs: appSharedPrefs)),
-        RepositoryProvider(create: (_) => dio)
+        RepositoryProvider(create: (_) => dio),
+        RepositoryProvider(create: (_) => serversRepository)
       ],
       child: MultiBlocProvider(
           providers: [
+            BlocProvider(create: (context) => memberCubit, lazy: false,),
+            BlocProvider(create: (context) => branchCubit, lazy: false,),
             BlocProvider(create: (context) => appCubit, lazy: false,),
             BlocProvider(create: (context) => localeCubit),
             BlocProvider(create: (context) => themeBloc),
-            BlocProvider(create: (context) => connectivityBloc)
+            BlocProvider(create: (context) => TutorialBloc(tutorialRepository: context.read<TutorialRepository>())),
+            BlocProvider(create: (context) => securityBloc),
+            BlocProvider(create: (context) => connectivityBloc),
+            BlocProvider(create: (context) => serversBloc)
           ],
-          child: App(localeCubit: localeCubit, themeBloc: themeBloc, dio: dio, rootConnectivityBloc: connectivityBloc, rootLocalization: rootLocalization, rootAppCubit: appCubit,))));
+          child: App(localeCubit: localeCubit, memberCubit: memberCubit, rootSecurityBloc: securityBloc, themeBloc: themeBloc, dio: dio, rootConnectivityBloc: connectivityBloc, rootLocalization: rootLocalization, rootAppCubit: appCubit,))));
 }
 
 class App extends StatefulWidget{
@@ -113,9 +135,11 @@ class App extends StatefulWidget{
   final AppLocalizations rootLocalization;
   final AppCubit rootAppCubit;
   final ThemeBloc themeBloc;
+  final SecurityBloc rootSecurityBloc;
   final ConnectivityBloc rootConnectivityBloc;
   final Dio dio;
-  const App({super.key, required this.dio, required this.localeCubit, required this.rootLocalization, required this.rootAppCubit, required this.rootConnectivityBloc, required this.themeBloc});
+  final MemberCubit memberCubit;
+  const App({super.key, required this.dio, required this.localeCubit, required this.rootLocalization, required this.rootAppCubit, required this.rootConnectivityBloc, required this.themeBloc, required this.rootSecurityBloc, required this.memberCubit});
 
   @override
   State<App> createState() => _AppState();
@@ -129,7 +153,7 @@ class _AppState extends State<App>{
   @override
   void initState() {
     super.initState();
-    router.addListener(_setStatusBarColor);
+    router.routerDelegate.addListener(_setStatusBarColor);
     _themeSubscription = widget.themeBloc.stream.listen((_) {
       _setStatusBarColor();
     });
@@ -153,7 +177,7 @@ class _AppState extends State<App>{
     super.dispose();
   }
 
-  bool get isRouteWithoutNavbar => router.location == '/' || router.location == RouteNames.signIn || router.location == RouteNames.resetPassword;
+  bool get isRouteWithoutNavbar => router.routeInformationProvider.value.uri.path == '/' || router.routeInformationProvider.value.uri.path == RouteNames.signInEmail || router.routeInformationProvider.value.uri.path == RouteNames.resetPassword;
 
   void _setStatusBarColor(){
     Future.delayed(const Duration(milliseconds: 300)).then((value){
@@ -194,52 +218,109 @@ class _AppState extends State<App>{
         _setStatusBarColor();
       }
     },
-    child: BlocBuilder<LocaleCubit,String>(
-      bloc: widget.localeCubit,
-      builder: (context,state) {
-        return CupertinoApp.router(
-            restorationScopeId: 'app',
-            routerConfig: router,
-            localizationsDelegates: const [
-              DefaultMaterialLocalizations.delegate,
-              GlobalMaterialLocalizations.delegate,
-              GlobalWidgetsLocalizations.delegate,
-              GlobalCupertinoLocalizations.delegate,
-              AppLocalizations.delegate,
-            ],
-          theme: const CupertinoThemeData(
-            brightness: Brightness.dark,
-            primaryColor: Colors.white,
-            textTheme: CupertinoTextThemeData(
-                textStyle: TextStyle(fontFamily: 'OpenSans')),
-          ),
-            locale: Locale(state),
-            supportedLocales: L10n.all,
-            title: 'Freeclimber',
-            debugShowCheckedModeBanner: false,
-            builder: (context,w){
-              return Theme(
-                data: ThemeData(
-                  fontFamily: 'OpenSans'
-                ),
-                child: OKToast(
-                  animationBuilder: const OffsetAnimationBuilder(maxOffsetY: -100),
-                    position: ToastPosition.top,
-                    child: Overlay(
-                      initialEntries:
-                      [
-                        OverlayEntry(builder: (context) => LoaderOverlay(
-                            overlayOpacity: 0,
-                            overlayColor: KColors.transparent,
-                            useDefaultLoading: false,
-                            overlayWidget: const OverlayLoadingWidget(),
-                            child: w!)),],
-                    )),
-              );
+    child: BlocListener<SecurityBloc, SecurityState>(
+    listener: (context, state) {
+      state.maybeMap(
+          initialized: (state) async{
+            final pushMessagesEnabled = !state.pushMessages;
+            if(pushMessagesEnabled){
+              await OneSignal.User.pushSubscription.optIn();
+            }else{
+              await OneSignal.User.pushSubscription.optOut();
+            }
+          },
+          failed: (_){
+            context.read<MemberCubit>().logout();
+          },
+          orElse: (){
+      });
+    },
+    child: BlocListener<MemberCubit, MemberState>(
+        listener: (context, state) {
+          if(state.status == MemberStateStatus.registrationSuccess){
+            if(state.memberRegistrationCode != null){
+              router.go(RouteNames.registrationComplete);
+            }else{
+              router.go(RouteNames.accountCreated);
+            }
+          }else if(state.status == MemberStateStatus.updateOperationSuccess){
+            AppToast.instance.showSuccess(context, widget.rootLocalization.success);
+          }else if(state.status == MemberStateStatus.unauthenticated){
+            router.go(RouteNames.signInEmail);
+          }else if(state.status == MemberStateStatus.error){
+            if(state.error!.localizationCode == 'socket-exception'){
+              AppToast.instance.showError(context, widget.rootLocalization.network_error);
+              return;
+            }else if(state.error!.localizationCode == 'timeout'){
+              return;
+            }
+            AppToast.instance.showError(context, state.error?.message ?? 'Unpredictable error');
+          }
+        },
+            listenWhen: (oldState,newState){
+              if(oldState.status == MemberStateStatus.error){
+                return false;
+              }else if(oldState.status == MemberStateStatus.updateOperationSuccess
+                  || oldState.status == MemberStateStatus.memberCardDownloaded && newState.status == MemberStateStatus.authenticated){
+                return false;
+              }else if(oldState.status == MemberStateStatus.authenticated && newState.status == MemberStateStatus.authenticated){
+                return false;
+              }
+              return true;
             },
-          );
-      },
-    ),
+        child: BlocBuilder<LocaleCubit,String>(
+          bloc: widget.localeCubit,
+          builder: (context,state) {
+            return CupertinoApp.router(
+                restorationScopeId: 'app',
+                routerConfig: router,
+                localizationsDelegates: const [
+                  DefaultMaterialLocalizations.delegate,
+                  GlobalMaterialLocalizations.delegate,
+                  GlobalWidgetsLocalizations.delegate,
+                  GlobalCupertinoLocalizations.delegate,
+                  AppLocalizations.delegate,
+                ],
+              theme: const CupertinoThemeData(
+                brightness: Brightness.dark,
+                primaryColor: Colors.white,
+                textTheme: CupertinoTextThemeData(
+                    textStyle: TextStyle(fontFamily: 'OpenSans')),
+              ),
+                locale: Locale(state),
+                supportedLocales: L10n.all,
+                title: 'Freeclimber',
+                debugShowCheckedModeBanner: false,
+                builder: (context,w){
+                  return Theme(
+                    data: ThemeData(
+                      fontFamily: 'OpenSans'
+                    ),
+                    child: OKToast(
+                      animationBuilder: const OffsetAnimationBuilder(maxOffsetY: -100),
+                        position: ToastPosition.top,
+                        child: Overlay(
+                          initialEntries:
+                          [
+                            OverlayEntry(builder: (context) => AppLifecycleOverlay(
+                              child: LoaderOverlay(
+                                  overlayColor: KColors.transparent,
+                                  useDefaultLoading: false,
+                                  overlayWidgetBuilder: (context) => const OverlayLoadingWidget(),
+                                  child: LockScreenOverlay(
+                                    appCubit: widget.rootAppCubit,
+                                      securityBloc: widget.rootSecurityBloc,
+                                      memberCubit: widget.memberCubit,
+                                      child: w!))
+                            )),],
+                        )),
+                  );
+                },
+              );
+          },
+        ),
+),
+),
 );
   }
 }
@@ -252,17 +333,20 @@ class HomePage extends StatefulWidget {
 }
 
 class _HomePageState extends State<HomePage> {
-  final bool _loaded = false;
+  bool _loaded = false;
 
   @override
   void initState() {
     super.initState();
+    _init();
     Future.delayed(Duration.zero, (){
       if(context.read<AppCubit>().defaultBranch != null){
         router.push(RouteNames.signUp);
       }
     });
   }
+
+  Future<void> _init() => context.read<MemberCubit>().fetchAuthenticationSession();
 
   @override
   void didChangeDependencies() {
@@ -271,6 +355,32 @@ class _HomePageState extends State<HomePage> {
 
   @override
   Widget build(BuildContext context) {
-    return _loaded ? const SignInScreen() : const SplashScreen();
+    return BlocListener<AppCubit, AppState>(
+    listener: (context, state) async{
+      if(state.status == AppStateStatus.loaded && context.read<MemberCubit>().state.status == MemberStateStatus.unauthenticated){
+        setState(() {
+          _loaded = true;
+        });
+      }else if(state.status == AppStateStatus.error){
+        setState(() {
+          _loaded = true;
+        });
+      }
+    },
+    child: BlocListener<MemberCubit, MemberState>(
+        listener: (context,state){
+          if(state.status != MemberStateStatus.loading){
+            context.loaderOverlay.hide();
+          } else if(state.status == MemberStateStatus.error){
+            if(state.error!.localizationCode == 'timeout'){
+              return;
+            }
+          }else if(state.status == MemberStateStatus.loading){
+            context.loaderOverlay.show();
+          }
+        },
+      child: _loaded ? const CreateServerScreen() : const SplashScreen(),
+      ),
+);
   }
 }
