@@ -6,8 +6,8 @@ part of 'app_documents.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-_$_AppDocuments _$$_AppDocumentsFromJson(Map<String, dynamic> json) =>
-    _$_AppDocuments(
+_$AppDocumentsImpl _$$AppDocumentsImplFromJson(Map<String, dynamic> json) =>
+    _$AppDocumentsImpl(
       privacy: json['privacy'] as String,
       branchPrivacy: json['branch_privacy'] as String?,
       imprint: json['imprint'] as String,
@@ -15,7 +15,7 @@ _$_AppDocuments _$$_AppDocumentsFromJson(Map<String, dynamic> json) =>
       branchTerms: json['branch_terms'] as String?,
     );
 
-Map<String, dynamic> _$$_AppDocumentsToJson(_$_AppDocuments instance) =>
+Map<String, dynamic> _$$AppDocumentsImplToJson(_$AppDocumentsImpl instance) =>
     <String, dynamic>{
       'privacy': instance.privacy,
       'branch_privacy': instance.branchPrivacy,
