@@ -6,12 +6,13 @@ part of 'color_set.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-_$_ColorSet _$$_ColorSetFromJson(Map<String, dynamic> json) => _$_ColorSet(
+_$ColorSetImpl _$$ColorSetImplFromJson(Map<String, dynamic> json) =>
+    _$ColorSetImpl(
       light: ThemeColors.fromJson(json['light'] as Map<String, dynamic>),
       dark: ThemeColors.fromJson(json['dark'] as Map<String, dynamic>),
     );
 
-Map<String, dynamic> _$$_ColorSetToJson(_$_ColorSet instance) =>
+Map<String, dynamic> _$$ColorSetImplToJson(_$ColorSetImpl instance) =>
     <String, dynamic>{
       'light': instance.light,
       'dark': instance.dark,
