@@ -1,17 +1,30 @@
 
 import 'dart:typed_data';
 
-import 'package:freeclimbers_employee/main.dart';
-import 'package:freeclimbers_employee/ui/screens/auth/account_created.dart';
-import 'package:freeclimbers_employee/ui/screens/auth/reset_password.dart';
-import 'package:freeclimbers_employee/ui/screens/auth/sign_in.dart';
-import 'package:freeclimbers_employee/ui/screens/auth/sign_up.dart';
-import 'package:freeclimbers_employee/ui/screens/home/settings/change_password_screen.dart';
-import 'package:freeclimbers_employee/ui/screens/home/settings/settings_screen.dart';
-import 'package:freeclimbers_employee/ui/screens/home/support/contact_us_screen.dart';
-import 'package:freeclimbers_employee/ui/screens/home/support/license_screen.dart';
-import 'package:freeclimbers_employee/ui/screens/home/support/support_screen.dart';
-import 'package:freeclimbers_employee/ui/screens/home/support/web_document_screen.dart';
+import 'package:climbers/blocs/extend_profile_bloc/extend_profile_bloc.dart';
+import 'package:climbers/main.dart';
+import 'package:climbers/ui/screens/auth/account_created.dart';
+import 'package:climbers/ui/screens/auth/create_server.dart';
+import 'package:climbers/ui/screens/auth/registration_complete.dart';
+import 'package:climbers/ui/screens/auth/reset_password.dart';
+import 'package:climbers/ui/screens/auth/sign_in_email.dart';
+import 'package:climbers/ui/screens/auth/sign_in_password.dart';
+import 'package:climbers/ui/screens/auth/sign_up.dart';
+import 'package:climbers/ui/screens/home/branch_details/branch_details_screen.dart';
+import 'package:climbers/ui/screens/home/branches/branches_screen.dart';
+import 'package:climbers/ui/screens/home/card/card_screen.dart';
+import 'package:climbers/ui/screens/home/extend_profile/extend_profile_screen.dart';
+import 'package:climbers/ui/screens/home/home.dart';
+import 'package:climbers/ui/screens/home/news/news_screen.dart';
+import 'package:climbers/ui/screens/home/profile/edit_profile_screen.dart';
+import 'package:climbers/ui/screens/home/profile/profile_screen.dart';
+import 'package:climbers/ui/screens/home/registration_code/registration_code_screen.dart';
+import 'package:climbers/ui/screens/home/settings/change_password_screen.dart';
+import 'package:climbers/ui/screens/home/settings/settings_screen.dart';
+import 'package:climbers/ui/screens/home/support/contact_us_screen.dart';
+import 'package:climbers/ui/screens/home/support/license_screen.dart';
+import 'package:climbers/ui/screens/home/support/support_screen.dart';
+import 'package:climbers/ui/screens/home/support/web_document_screen.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:go_router/go_router.dart';
 
@@ -20,29 +33,104 @@ import 'consts/web_documents_enums.dart';
 class RouteNames{
   ///Supports [extra] bool variable.
   ///True if you want to show animation
-  static String signIn = '/sign_in';
+  static String signInEmail = '/sign_in_email';
+
+  ///Requires [extra] variable of type string, which is email from previous step
+  static String signInPassword = '/sign_in_password';
   static String main = '/';
   static String signUp = '/sign_up';
+  ///Supports [extra] variable of type string, which is email from login form
   static String resetPassword = '/reset_password';
   static String accountCreated = '/account_created';
 
+  ///Supports [extra] variable of type int where 1 is no transition animation, 2 is vertical, 3 is horizontal.
+  static String home = '/home';
+  static String branch = '/branch';
   ///Supports [extra] variable of type [WebDocumentRouteExtra] json
   static String webDocument(String lang, WebDocumentRoutes route, [int? branchId]) => '/web_document/$lang/${route.route}${branchId == null ? '' : '?branch_id=$branchId'}';
   ///Supports [extra] variable of type [WebBranchDocumentRouteExtra] json
   static String webBranchDocument(String lang, WebBranchDocumentRoutes route) => '/web_branch_document/$lang/${route.route}';
   static String license = '/license';
+  static String card = '/card';
+  static String profile = '/profile';
   static String settings = '/settings';
   static String help = '/help';
-  static String contactUs = '/contact_us';
+  static String registrationCode = '/registration_code';
+  static String news = '/news';
+  static String branchDetails = '/branch_details';
 
+  static String editProfile = '/edit_profile';
   static String changePassword = '/change_password';
 
+  ///Attach [extra] of type ExtendProfileBloc
+  static String extendProfile = '/extend_profile';
+
   static String registrationComplete = '/registration_complete';
+
+  static String contactUs = '/contact_us';
+
+  static String createServer = '/create_server';
 }
 
 final GoRouter router = GoRouter(
   restorationScopeId: 'router',
+    initialLocation: RouteNames.createServer,
     routes: <GoRoute>[
+      GoRoute(
+          path: RouteNames.registrationCode,
+          pageBuilder: (context, state) => CustomTransitionPage<void>(
+            opaque: true,
+            key: state.pageKey,
+            child: const RegistrationCodeScreen(),
+            transitionsBuilder: (context, animation, secondaryAnimation, child) =>
+                SlideTransition(
+                    position: Tween<Offset>(
+                        begin: const Offset(0,1),
+                        end: Offset.zero
+                    ).animate(animation),
+                    child: child),
+          ),
+          builder: (context,state){
+            return const RegistrationCodeScreen();
+          }
+      ),
+      GoRoute(
+          path: RouteNames.createServer,
+          pageBuilder: (context, state) => CustomTransitionPage<void>(
+            transitionDuration: const Duration(milliseconds: 400),
+            opaque: true,
+            key: state.pageKey,
+            child: const CreateServerScreen(),
+            transitionsBuilder: (context, animation, secondaryAnimation, child) =>
+                SlideTransition(
+                    position: Tween<Offset>(
+                        begin: const Offset(0, 1),
+                        end: Offset.zero
+                    ).animate(animation),
+                    child: child),
+          ),
+          builder: (context,state){
+            return const CreateServerScreen();
+          }
+      ),
+      GoRoute(
+          path: RouteNames.extendProfile,
+          pageBuilder: (context, state) => CustomTransitionPage<void>(
+            opaque: true,
+            key: state.pageKey,
+            child: ExtendProfileScreen(extendProfileBloc: state.extra as ExtendProfileBloc),
+            transitionsBuilder: (context, animation, secondaryAnimation, child) =>
+                SlideTransition(
+                    position: Tween<Offset>(
+                        begin: const Offset(1,0),
+                        end: Offset.zero
+                    ).animate(animation),
+                    child: child),
+          ),
+          builder: (context,state){
+            return ExtendProfileScreen(extendProfileBloc: state.extra as ExtendProfileBloc,);
+          }
+      ),
       GoRoute(
           path: RouteNames.contactUs,
           pageBuilder: (context, state) => CustomTransitionPage<void>(
@@ -59,6 +147,60 @@ final GoRouter router = GoRouter(
           ),
           builder: (context,state){
             return ContactUsScreen();
+          }
+      ),
+      GoRoute(
+          path: RouteNames.registrationComplete,
+          pageBuilder: (context, state) => CustomTransitionPage<void>(
+            opaque: true,
+            key: state.pageKey,
+            child: const RegistrationComplete(isAppClip: false),
+            transitionsBuilder: (context, animation, secondaryAnimation, child) =>
+                SlideTransition(
+                    position: Tween<Offset>(
+                        begin: const Offset(1,0),
+                        end: Offset.zero
+                    ).animate(animation),
+                    child: child),
+          ),
+          builder: (context,state){
+            return const RegistrationComplete(isAppClip: false,);
+          }
+      ),
+      GoRoute(
+          path: RouteNames.branchDetails,
+          pageBuilder: (context, state) => CustomTransitionPage<void>(
+            opaque: true,
+            key: state.pageKey,
+            child: const BranchDetailsScreen(),
+            transitionsBuilder: (context, animation, secondaryAnimation, child) =>
+                SlideTransition(
+                    position: Tween<Offset>(
+                        begin: const Offset(1,0),
+                        end: Offset.zero
+                    ).animate(animation),
+                    child: child),
+          ),
+          builder: (context,state){
+            return const BranchDetailsScreen();
+          }
+      ),
+      GoRoute(
+          path: RouteNames.news,
+          pageBuilder: (context, state) => CustomTransitionPage<void>(
+            opaque: true,
+            key: state.pageKey,
+            child: const NewsScreen(),
+            transitionsBuilder: (context, animation, secondaryAnimation, child) =>
+                SlideTransition(
+                    position: Tween<Offset>(
+                        begin: const Offset(1,0),
+                        end: Offset.zero
+                    ).animate(animation),
+                    child: child),
+          ),
+          builder: (context,state){
+            return const NewsScreen();
           }
       ),
       GoRoute(
@@ -85,7 +227,7 @@ final GoRouter router = GoRouter(
           pageBuilder: (context, state) => CustomTransitionPage<void>(
             opaque: true,
             key: state.pageKey,
-            child: WebDocumentScreen(locale: state.params["lang"] as String,route: state.params["route"] as String, fromRequest: true, content: (state.extra as Map)["content"], overrideBranch: (state.extra as Map)["overrideIcon"] == null ? null : (state.extra as Map)["overrideIcon"]!,),
+            child: WebDocumentScreen(locale: state.uri.queryParameters["lang"] as String,route: state.uri.queryParameters["route"] as String, fromRequest: true, content: (state.extra as Map)["content"], overrideBranch: (state.extra as Map)["overrideIcon"] == null ? null : (state.extra as Map)["overrideIcon"]!,),
             transitionsBuilder: (context, animation, secondaryAnimation, child) =>
                 SlideTransition(
                     position: Tween<Offset>(
@@ -95,7 +237,7 @@ final GoRouter router = GoRouter(
                     child: child),
           ),
           builder: (context,state){
-            return WebDocumentScreen(locale: state.params["lang"] as String,route: state.params["route"] as String, fromRequest: true, content: (state.extra as Map)["content"], overrideBranch: (state.extra as Map)["overrideIcon"] == null ? null : (state.extra as Map)["overrideIcon"]!);
+            return WebDocumentScreen(locale: state.uri.queryParameters["lang"] as String,route: state.uri.queryParameters["route"] as String, fromRequest: true, content: (state.extra as Map)["content"], overrideBranch: (state.extra as Map)["overrideIcon"] == null ? null : (state.extra as Map)["overrideIcon"]!);
           }
       ),
       GoRoute(
@@ -104,7 +246,7 @@ final GoRouter router = GoRouter(
           pageBuilder: (context, state) => CustomTransitionPage<void>(
             opaque: true,
             key: state.pageKey,
-            child: WebDocumentScreen(locale: state.params["lang"] as String,route: state.params["route"] as String, overrideBranch: state.extra as Uint8List?,),
+            child: WebDocumentScreen(locale: state.uri.queryParameters["lang"] as String,route: state.uri.queryParameters["route"] as String, overrideBranch: state.extra as Uint8List?,),
             transitionsBuilder: (context, animation, secondaryAnimation, child) =>
                 SlideTransition(
                     position: Tween<Offset>(
@@ -114,7 +256,43 @@ final GoRouter router = GoRouter(
                     child: child),
           ),
           builder: (context,state){
-            return WebDocumentScreen(locale: state.params["lang"] as String,route: state.params["route"] as String, overrideBranch: state.extra as Uint8List?);
+            return WebDocumentScreen(locale: state.uri.queryParameters["lang"] as String,route: state.uri.queryParameters["route"] as String, overrideBranch: state.extra as Uint8List?);
+          }
+      ),
+      GoRoute(
+          path: RouteNames.branch,
+          pageBuilder: (context, state) => CustomTransitionPage<void>(
+            opaque: true,
+            key: state.pageKey,
+            child: const BranchesScreen(),
+            transitionsBuilder: (context, animation, secondaryAnimation, child) =>
+                SlideTransition(
+                    position: Tween<Offset>(
+                        begin: const Offset(1,0),
+                        end: Offset.zero
+                    ).animate(animation),
+                    child: child),
+          ),
+          builder: (context,state){
+            return const BranchesScreen();
+          }
+      ),
+      GoRoute(
+          path: RouteNames.card,
+          pageBuilder: (context, state) => CustomTransitionPage<void>(
+            opaque: true,
+            key: state.pageKey,
+            child: const CardScreen(),
+            transitionsBuilder: (context, animation, secondaryAnimation, child) =>
+                SlideTransition(
+                    position: Tween<Offset>(
+                        begin: const Offset(0,1),
+                        end: Offset.zero
+                    ).animate(animation),
+                    child: child),
+          ),
+          builder: (context,state){
+            return const CardScreen();
           }
       ),
       GoRoute(
@@ -172,28 +350,100 @@ final GoRouter router = GoRouter(
           }
       ),
       GoRoute(
-        path: RouteNames.signIn,
+          path: RouteNames.editProfile,
+          pageBuilder: (context, state) => CustomTransitionPage<void>(
+            opaque: true,
+            key: state.pageKey,
+            child: const EditProfileScreen(),
+            transitionsBuilder: (context, animation, secondaryAnimation, child) =>
+                SlideTransition(
+                    position: Tween<Offset>(
+                        begin: const Offset(1,0),
+                        end: Offset.zero
+                    ).animate(animation),
+                    child: child),
+          ),
+          builder: (context,state){
+            return const EditProfileScreen();
+          }
+      ),
+      GoRoute(
+          path: RouteNames.profile,
+          pageBuilder: (context, state) => CustomTransitionPage<void>(
+            opaque: true,
+            key: state.pageKey,
+            child: const ProfileScreen(),
+            transitionsBuilder: (context, animation, secondaryAnimation, child) =>
+                SlideTransition(
+                    position: Tween<Offset>(
+                        begin: const Offset(1,0),
+                        end: Offset.zero
+                    ).animate(animation),
+                    child: child),
+          ),
+          builder: (context,state){
+            return const ProfileScreen();
+          }
+      ),
+      GoRoute(
+        path: RouteNames.signInEmail,
           pageBuilder: (context, state) => CustomTransitionPage<void>(
             transitionDuration: const Duration(milliseconds: 400),
             opaque: true,
             key: state.pageKey,
-            child: const SignInScreen(),
+            child: const SignInEmailScreen(),
             transitionsBuilder: (context, animation, secondaryAnimation, child) =>
-                Stack(
-                  children: [
-                    SignInScreen(),
-                    SlideTransition(
-                        position: Tween<Offset>(
-                          begin: const Offset(0, - 1),
-                          end: Offset.zero
-                        ).animate(animation),
-                        child: child),
-                  ],
-                ),
+                child,
           ),
         builder: (context,state){
-          return const SignInScreen();
+          return const SignInEmailScreen();
         }
+      ),
+      GoRoute(
+          path: RouteNames.signInPassword,
+          pageBuilder: (context, state) => CustomTransitionPage<void>(
+            transitionDuration: const Duration(milliseconds: 400),
+            opaque: true,
+            key: state.pageKey,
+            child: SignInPasswordScreen(email: state.extra as String,),
+            transitionsBuilder: (context, animation, secondaryAnimation, child) =>
+            child,
+          ),
+          builder: (context,state){
+            return SignInPasswordScreen(email: state.extra as String,);
+          }
+      ),
+      GoRoute(
+          path: RouteNames.home,
+          pageBuilder: (context, state) => CustomTransitionPage<void>(
+            restorationId: RouteNames.home,
+            transitionDuration: const Duration(milliseconds: 400),
+            opaque: true,
+            key: state.pageKey,
+            child: const Home(),
+            transitionsBuilder: (context, animation, secondaryAnimation, child) =>
+                state.extra != null && state.extra == 1 ? FadeTransition(
+                    opacity: animation,
+                    child: child) : state.extra == null || state.extra == 2 ? SlideTransition(
+                    position: Tween<Offset>(
+                        begin: const Offset(1, 0),
+                        end: Offset.zero
+                    ).animate(animation),
+                    child: child) : Stack(
+                      children: [
+                        const Home(),
+                        SlideTransition(
+                            position: Tween<Offset>(
+                                begin: Offset.zero,
+                                end: const Offset(0, -1),
+                            ).animate(animation),
+                            child: child),
+                      ],
+                    ),
+          ),
+          builder: (context,state){
+            return const Home();
+          }
       ),
       GoRoute(
           path: RouteNames.resetPassword,
@@ -201,7 +451,7 @@ final GoRouter router = GoRouter(
             transitionDuration: const Duration(milliseconds: 400),
             opaque: true,
             key: state.pageKey,
-            child: const ResetPassword(),
+            child: ResetPassword(email: state.extra as String?,),
             transitionsBuilder: (context, animation, secondaryAnimation, child) =>
                 SlideTransition(
                     position: Tween<Offset>(
@@ -211,7 +461,7 @@ final GoRouter router = GoRouter(
                     child: child),
           ),
           builder: (context,state){
-            return const ResetPassword();
+            return ResetPassword(email: state.extra as String?,);
           }
       ),
       GoRoute(
