@@ -1,6 +1,6 @@
 import 'dart:developer';
 
-import 'package:freeclimbers_employee/services/membercard/membercard_service_contract.dart';
+import 'package:climbers/services/membercard/membercard_service_contract.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_wallet/flutter_wallet.dart';
 
