@@ -1,13 +1,14 @@
 
-import 'package:freeclimbers_employee/extensions/colors.dart';
-import 'package:freeclimbers_employee/ui/widgets/buttons/app_elevated_button.dart';
-import 'package:freeclimbers_employee/ui/widgets/connectivity_badge/connectivity_badge.dart';
-import 'package:freeclimbers_employee/utils/validators.dart';
+import 'package:climbers/extensions/colors.dart';
+import 'package:climbers/ui/widgets/buttons/app_elevated_button.dart';
+import 'package:climbers/ui/widgets/connectivity_badge/connectivity_badge.dart';
+import 'package:climbers/utils/validators.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:go_router/go_router.dart';
 import 'package:loader_overlay/loader_overlay.dart';
+import '../../../blocs/member_cubit/member_cubit.dart';
 import '../../../utils/toasts.dart';
 import '../../widgets/buttons/app_outlined_button.dart';
 
@@ -15,7 +16,8 @@ import '../../widgets/text_fields/app_plain_text_field.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 
 class ResetPassword extends StatefulWidget {
-  const ResetPassword({Key? key}) : super(key: key);
+  final String? email;
+  const ResetPassword({Key? key, this.email}) : super(key: key);
 
   @override
   State<ResetPassword> createState() => _ResetPasswordState();
@@ -23,7 +25,7 @@ class ResetPassword extends StatefulWidget {
 
 class _ResetPasswordState extends State<ResetPassword> {
   final _formKey = GlobalKey<FormState>();
-  final _emailController = TextEditingController();
+  late final _emailController = TextEditingController(text: widget.email ?? '');
   String _errorTag = '';
 
   @override
@@ -34,61 +36,82 @@ class _ResetPasswordState extends State<ResetPassword> {
   @override
   Widget build(BuildContext context) {
     return CupertinoPageScaffold(
-        backgroundColor: context.colors!.basePrimaryBack,
+        backgroundColor: context.colors.basePrimaryBack,
         child: ConnectivityBadge(
-          child: Form(
-            key: _formKey,
-            child: Align(
-              alignment: Alignment.topCenter,
-              child: Padding(
-                padding: EdgeInsets.symmetric(horizontal: MediaQuery.of(context).size.width * 0.06),
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Spacer(flex: 2,),
-                    Image.asset('assets/images/logo.png', width: 100, height: 100,),
-                    const SizedBox(height: 20,),
-                    Text(AppLocalizations.of(context)!.reset_password, style: TextStyle(color: context.colors!.fieldNormalText, fontSize: 32, fontWeight: FontWeight.w700),),
-                    const SizedBox(height: 24,),
-                    Text(AppLocalizations.of(context)!.reset_password_hint, style: TextStyle(fontSize: 16, fontWeight: FontWeight.w600, color: context.colors!.fieldNormalText),),
-                    const SizedBox(height: 24),
-                    AppPlanTextField(title: AppLocalizations.of(context)!.email,
-                      whiteTitle: false,
-                      currentTags: [_errorTag],
-                      prefixWidget: Icon(FontAwesomeIcons.solidEnvelope, color: context.colors!.fieldNormalText, size: 20,),
-                      mandatority: 0,
-                      textCapitalization: TextCapitalization.none,
-                      tagForError: 'password_reset',
-                      controller: _emailController, validator: AppValidators.email,),
-                    SizedBox(height: 32,),
-                    AppElevatedButton(
-                      title: AppLocalizations.of(context)!.reset_password, onPressed: (){
-                        _errorTag = '';
-                        if(_emailController.text.trim().isEmpty){
-                          setState((){
-                            _errorTag = 'password_reset';
-                          });
-                          Future.delayed(Duration(milliseconds: 100)).then((_){
-                            _formKey.currentState!.validate();
-                          });
-                          AppToast.instance.showError(context, AppLocalizations.of(context)!.reset_password_empty_error);
-                          return;
-                        }else if(AppValidators.email(_emailController.text) != null){
-                          setState((){
-                            _errorTag = 'password_reset';
-                          });
-                          Future.delayed(Duration(milliseconds: 100)).then((_){
-                            _formKey.currentState!.validate();
-                          });
-                          AppToast.instance.showError(context, AppLocalizations.of(context)!.reset_password_email_not_valid_error);
-                          return;
-                        }
-                        //context.loaderOverlay.show();
-                    }),
-                    const SizedBox(height: 32,),
-                    AppOutlinedButton(title: AppLocalizations.of(context)!.back, onPressed: () => context.pop()),
-                    const Spacer(flex: 2,)
-                  ],
+          child: BlocListener<MemberCubit,MemberState>(
+            listener: (context,state){
+              if(state.status == MemberStateStatus.error){
+                if(state.error!.localizationCode == 'timeout'){
+                  return;
+                }
+                context.loaderOverlay.hide();
+                setState((){
+                  _errorTag = 'password_reset';
+                });
+                Future.delayed(Duration(milliseconds: 100)).then((_){
+                  _formKey.currentState!.validate();
+                });
+              }else if(state.status == MemberStateStatus.sentEmail){
+                context.loaderOverlay.hide();
+                AppToast.instance.showSuccess(context, AppLocalizations.of(context)!.verification_sent);
+                context.pop();
+              }
+            },
+            child: Form(
+              key: _formKey,
+              child: Align(
+                alignment: Alignment.topCenter,
+                child: Padding(
+                  padding: EdgeInsets.symmetric(horizontal: MediaQuery.of(context).size.width * 0.06),
+                  child: Column(
+                    crossAxisAlignment: CrossAxisAlignment.start,
+                    children: [
+                      Spacer(flex: 2,),
+                      Image.asset('assets/images/logo.png', width: 100, height: 100,),
+                      const SizedBox(height: 20,),
+                      Text(AppLocalizations.of(context)!.reset_password, style: TextStyle(color: context.colors.fieldNormalText, fontSize: 32, fontWeight: FontWeight.w700),),
+                      const SizedBox(height: 24,),
+                      Text(AppLocalizations.of(context)!.reset_password_hint, style: TextStyle(fontSize: 16, fontWeight: FontWeight.w600, color: context.colors.fieldNormalText),),
+                      const SizedBox(height: 24),
+                      AppPlanTextField(title: AppLocalizations.of(context)!.email,
+                        whiteTitle: false,
+                        currentTags: [_errorTag],
+                        prefixWidget: Icon(FontAwesomeIcons.solidEnvelope, color: context.colors.fieldNormalText, size: 20,),
+                        mandatority: 0,
+                        textCapitalization: TextCapitalization.none,
+                        tagForError: 'password_reset',
+                        controller: _emailController, validator: AppValidators.email,),
+                      SizedBox(height: 32,),
+                      AppElevatedButton(
+                        title: AppLocalizations.of(context)!.reset_password, onPressed: (){
+                          _errorTag = '';
+                          if(_emailController.text.trim().isEmpty){
+                            setState((){
+                              _errorTag = 'password_reset';
+                            });
+                            Future.delayed(Duration(milliseconds: 100)).then((_){
+                              _formKey.currentState!.validate();
+                            });
+                            AppToast.instance.showError(context, AppLocalizations.of(context)!.reset_password_empty_error);
+                            return;
+                          }else if(AppValidators.email(_emailController.text) != null){
+                            setState((){
+                              _errorTag = 'password_reset';
+                            });
+                            Future.delayed(Duration(milliseconds: 100)).then((_){
+                              _formKey.currentState!.validate();
+                            });
+                            AppToast.instance.showError(context, AppLocalizations.of(context)!.reset_password_email_not_valid_error);
+                            return;
+                          }
+                          context.loaderOverlay.show();
+                          context.read<MemberCubit>().resetPassword(email: _emailController.text);
+                      }),
+                      const SizedBox(height: 32,),
+                      AppOutlinedButton(title: AppLocalizations.of(context)!.back, onPressed: () => context.pop()),
+                      const Spacer(flex: 2,)
+                    ],
+                  ),
                 ),
               ),
             ),
