@@ -1,5 +1,5 @@
-import 'package:freeclimbers_employee/extensions/colors.dart';
-import 'package:freeclimbers_employee/utils/license.dart';
+import 'package:climbers/extensions/colors.dart';
+import 'package:climbers/utils/license.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:go_router/go_router.dart';
