@@ -1,5 +1,5 @@
-import 'package:freeclimbers_employee/consts/k_colors.dart';
-import 'package:freeclimbers_employee/extensions/colors.dart';
+import 'package:climbers/consts/k_colors.dart';
+import 'package:climbers/extensions/colors.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
