@@ -1,8 +1,8 @@
 
-import 'package:freeclimbers_employee/blocs/theme_bloc/theme_bloc.dart';
-import 'package:freeclimbers_employee/consts/k_colors.dart';
-import 'package:freeclimbers_employee/extensions/colors.dart';
-import 'package:freeclimbers_employee/utils/validators.dart';
+import 'package:climbers/blocs/theme_bloc/theme_bloc.dart';
+import 'package:climbers/consts/k_colors.dart';
+import 'package:climbers/extensions/colors.dart';
+import 'package:climbers/utils/validators.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
